requirements.txt
markdown-it-py
mdit-py-plugins
latex2mathml

